feat: Store-Ebene für Teilstand - Preise nullable, Vollständigkeits-Query
Ticket #128 (Wayfinder-Map #123, Spec #127): der Store kann eine Ablesung
mit teilweise fehlenden Werten anlegen/aktualisieren und weiß zuverlässig,
ob eine Periode vollständig ist.

- periods' 4 Preis-Spalten per Tabellen-Rebuild-Migration nullable
  (SQLite kennt kein ALTER COLUMN DROP NOT NULL), idempotent, gegen eine
  echte Bestands-DB verifiziert (Daten/FK-Bezüge bleiben byte-identisch).
- PeriodInput/LatestPeriod: die 4 Preis-Felder werden *float64 (nil =
  nicht erfasst). Readings/Personen dürfen jetzt sparse sein - ein
  fehlender Meter/eine fehlende Wohnung wird einfach nicht persistiert,
  statt einen Fehler zu werfen.
- UpdatePeriod überschreibt ein leeres/fehlendes Feld nicht mehr mit NULL,
  sondern lässt den vorhandenen Wert stehen - dieselbe Regel wie schon für
  Personen, jetzt auch für Preise/Monat/Readings.
- checkMonatNeighbors überspringt die Nachbar-Prüfung bei leerem Monat.
- Neue Vollständigkeits-Query PeriodComplete, rein abgeleitet (kein
  eigenes Flag).
- GetPeriodByID (verbrauch.go) liest einen Teilstand-Preis NULL-sicher als
  0 - Period füttert nur calc's Kosten-Berechnung, die nie auf eine
  unvollständige Periode läuft.
- Mechanischer Compile-Fix in allen anderen PeriodInput/LatestPeriod-
  Aufrufstellen (internal/web, internal/calc-Tests) - keine Verhaltens-
  änderung dort, das ist Ticket #129.

CONTEXT.md um den Begriff "Teilstand" ergänzt.

+++ internal/store/demoseed.go
@@ -156,11 +156,11 @@ func seedDemoPeriods(db *sql.DB, now time.Time) error {
 		in := PeriodInput{
 			ReadingDate:             date.Format("2006-01-02"),
 			Monat:                   monat,
-			Strompreis:              strompreis,
-			FrischwasserPreis:       frischwasserPreis,
-			AbwasserPreis:           abwasserPreis,
+			Strompreis:              Float64(strompreis),
+			FrischwasserPreis:       Float64(frischwasserPreis),
+			AbwasserPreis:           Float64(abwasserPreis),
 			HeizungWaermeGewichtung: 0.7,
-			EinspeisungPreis:        0.08,
+			EinspeisungPreis:        Float64(0.08),
 			Readings:                readingsCopy,
 			Personen:                map[int64]int64{1: 2, 2: personenWohnung2(i)},
 		}
